refactor(internal/argocd): Isola as operações de leitura em um pacote de status
Para melhorar a organização do código e seguir o Princípio da
Responsabilidade Única, esta refatoração separa as operações de
leitura (Get/List) das operações de escrita (Create/Delete) dentro
do pacote 'argocd'.

- Um novo arquivo  foi criado para centralizar todas as
  funções que buscam o estado de recursos do Argo CD.

- As funções , , ,
  ,  e  foram movidas
  para .

- Os arquivos , , e
  foram simplificados para conter apenas a lógica de escrita.

Esta alteração não introduz nenhuma mudança de funcionalidade, mas
torna a base de código mais limpa, mais fácil de navegar e mais
sustentável para futuras adições.
