+++ internal/argocd/project.go
@@ -78,41 +78,3 @@ func DeleteProject(ctx context.Context, serverAddr, authToken string, insecure b
 	}
 	return nil
 }
-
-// GetProject busca um projeto específico pelo nome.
-func GetProject(ctx context.Context, serverAddr, authToken string, insecure bool, projName string) (*v1alpha1.AppProject, error) {
-	apiClient, err := NewClient(serverAddr, authToken, insecure)
-	if err != nil {
-		return nil, err
-	}
-	projServiceCloser, projServiceClient, err := apiClient.NewProjectClient()
-	if err != nil {
-		return nil, fmt.Errorf("falha ao obter o cliente de projeto: %w", err)
-	}
-	defer projServiceCloser.Close()
-
-	p, err := projServiceClient.Get(ctx, &project.ProjectQuery{Name: projName})
-	if err != nil {
-		return nil, fmt.Errorf("falha ao buscar o projeto '%s': %w", projName, err)
-	}
-	return p, nil
-}
-
-// ListProjects busca todos os projetos registrados no Argo CD.
-func ListProjects(ctx context.Context, serverAddr, authToken string, insecure bool) ([]v1alpha1.AppProject, error) {
-	apiClient, err := NewClient(serverAddr, authToken, insecure)
-	if err != nil {
-		return nil, err
-	}
-	projServiceCloser, projServiceClient, err := apiClient.NewProjectClient()
-	if err != nil {
-		return nil, fmt.Errorf("falha ao obter o cliente de projeto: %w", err)
-	}
-	defer projServiceCloser.Close()
-
-	projectList, err := projServiceClient.List(ctx, &project.ProjectQuery{})
-	if err != nil {
-		return nil, fmt.Errorf("falha ao listar os projetos: %w", err)
-	}
-	return projectList.Items, nil
-}
