+++ internal/argocd/repository.go
@@ -77,46 +77,3 @@ func DeleteRepository(ctx context.Context, serverAddr, authToken string, insecur
 	}
 	return nil
 }
-
-// GetRepository busca um repositório específico pela URL.
-func GetRepository(ctx context.Context, serverAddr, authToken string, insecure bool, repoURL string) (*v1alpha1.Repository, error) {
-	apiClient, err := NewClient(serverAddr, authToken, insecure)
-	if err != nil {
-		return nil, err
-	}
-	repoServiceCloser, repoServiceClient, err := apiClient.NewRepoClient()
-	if err != nil {
-		return nil, fmt.Errorf("falha ao obter o cliente de repositório: %w", err)
-	}
-	defer repoServiceCloser.Close()
-
-	repo, err := repoServiceClient.Get(ctx, &repository.RepoQuery{Repo: repoURL})
-	if err != nil {
-		return nil, fmt.Errorf("falha ao buscar o repositório '%s': %w", repoURL, err)
-	}
-	return repo, nil
-}
-
-// ListRepositories busca todos os repositórios registrados no Argo CD.
-func ListRepositories(ctx context.Context, serverAddr, authToken string, insecure bool) ([]v1alpha1.Repository, error) {
-	apiClient, err := NewClient(serverAddr, authToken, insecure)
-	if err != nil {
-		return nil, err
-	}
-	repoServiceCloser, repoServiceClient, err := apiClient.NewRepoClient()
-	if err != nil {
-		return nil, fmt.Errorf("falha ao obter o cliente de repositório: %w", err)
-	}
-	defer repoServiceCloser.Close()
-
-	repoList, err := repoServiceClient.List(ctx, &repository.RepoQuery{})
-	if err != nil {
-		return nil, fmt.Errorf("falha ao listar os repositórios: %w", err)
-	}
-
-	repos := make([]v1alpha1.Repository, len(repoList.Items))
-	for i, item := range repoList.Items {
-		repos[i] = *item
-	}
-	return repos, nil
-}
